Move resources to v1
