+++ src/main/java/com/patreon/PatreonAPI.java
@@ -3,10 +3,10 @@ package com.patreon;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.github.jasminb.jsonapi.*;
-import com.patreon.resources.Campaign;
-import com.patreon.resources.Pledge;
-import com.patreon.resources.RequestUtil;
-import com.patreon.resources.User;
+import com.patreon.resources.v1.Campaign;
+import com.patreon.resources.v1.Pledge;
+import com.patreon.resources.v1.RequestUtil;
+import com.patreon.resources.v1.User;
 import com.patreon.resources.shared.BaseResource;
 import com.patreon.resources.shared.Field;
 import org.apache.http.NameValuePair;
@@ -17,7 +17,6 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.lang.reflect.Array;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.*;
