+++ src/main/java/com/patreon/resources/v1/Pledge.java
@@ -1,4 +1,4 @@
-package com.patreon.resources;
+package com.patreon.resources.v1;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.github.jasminb.jsonapi.annotations.Relationship;
@@ -6,10 +6,8 @@ import com.github.jasminb.jsonapi.annotations.Type;
 import com.patreon.resources.shared.BaseResource;
 import com.patreon.resources.shared.Field;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
 import java.util.stream.Collectors;
 
 @Type("pledge")
